--- src/main/java/caculator/domain/Operator.java
@@ -1,30 +1,0 @@
-package caculator.domain;
-
-import java.util.Arrays;
-import java.util.function.BiFunction;
-
-public enum Operator {
-    PLUS("+", (num1, num2) -> num1 + num2),
-    MINUS("-", (num1, num2) -> num1 - num2),
-    MULTIPLY("*", (num1, num2) -> num1 * num2),
-    DIVIDE("/", (num1, num2) -> num1 / num2);
-
-    private String operator;
-    private BiFunction<Integer, Integer, Integer> expression;
-
-    Operator(String operator, BiFunction<Integer, Integer, Integer> expression) {
-        this.operator = operator;
-        this.expression = expression;
-    }
-
-    public String getOperator() {
-        return operator;
-    }
-
-    public Integer calculate(int num1, int num2) {
-        return this.expression.apply(num1, num2);
-    }
-}
-
-
-//열거형에 값을입력할 경우: 생성자를 통해 해당 값에 넣어준다.
